+++ R/predict.BayesSurvive.R
@@ -27,10 +27,12 @@
 #' Cox model
 #' @param \dots not used
 #'
-#' @return A list object including seven components with the first compoment as
-#' the specified argument \code{type}. The other components of the list are
-#' "se", "band", "type", "diag", "baseline" and "times", see function
-#' \code{riskRegression::predictCox} for details
+#' @return A list object with 5 components if \code{type="brier"} including
+#'"model", "times", "Brier", "IBS" and "IPA" (Index of Prediction Accuracy), 
+#' otherwise  a list of 7 components with the first component as
+#' the specified argument \code{type} and "se", "band", "type", "diag", 
+#' "baseline" and "times", see function \code{riskRegression::predictCox} for 
+#' details
 #'
 #' @examples
 #'
@@ -95,6 +97,7 @@ predict.BayesSurvive <- function(object, survObj.new, type = "brier",
     beta_m <- object$output$beta.margin # colMeans(betas)
     betas <- object$output$beta.p[-(1:(object$input$burnin / object$input$thin + 1)), ]
     survObj <- object$output$survObj
+    survObj$lp.all <- survObj$lp.all[, -1]
   } else {
     beta_m <- object$output$beta.margin[[subgroup]]
     survObj.new <- survObj.new[[subgroup]]
@@ -103,7 +106,7 @@ predict.BayesSurvive <- function(object, survObj.new, type = "brier",
     survObj$lp.all <- survObj$lp.all[, -(1:(object$input$burnin / object$input$thin + 1))]
   }
 
-  ibs <- data.frame("Null model" = NA, "Bayesian Cox" = NA)
+  ibs <- data.frame("Null model" = rep(NA, 3), "Bayesian Cox" = rep(NA, 3))
 
   if (is.null(times)) {
     times <- sort(unique(survObj.new$t))
@@ -209,18 +212,20 @@ predict.BayesSurvive <- function(object, survObj.new, type = "brier",
         Brier <- Score(list("Bayesian Cox" = model_train),
           formula = Surv(time, status) ~ 1,
           data = data_test, conf.int = FALSE,
-          metrics = "brier", summary = "ibs",
-          times = times
+          metrics = "brier", summary = c("ibs", "ipa"),
+          null.model = TRUE, times = times
         )$Brier$score
+        
         # Brier <- BrierScore[BrierScore$model != "Null model", ]
-        # extract IBS for Null model and the Bayesian Cox model
-        ibs[1, ] <- Brier$IBS[c(length(times), length(times) * 2)]
+        # extract scores for Null model and the Bayesian Cox model
+        ibs[1, ] <- Brier$Brier[c(length(times), length(times) * 2)]
+        ibs[2, ] <- Brier$IBS[c(length(times), length(times) * 2)]
+        ibs[3, ] <- Brier$IPA[c(length(times), length(times) * 2)]
       }
     } else {
       # obtain linear predictors corresponding to MCMC estimates
       # lp_all_train <- survObj_train[, -c(1:2)] %*% t(betas)
       lp_all_test <- survObj.new$X %*% t(betas)
-      Brier <- matrix(0, nrow = length(times) * 2, ncol = 2)
       data_train <- data.frame(
         time = survObj$t,
         status = survObj$di,
@@ -236,13 +241,17 @@ predict.BayesSurvive <- function(object, survObj.new, type = "brier",
         y = TRUE, x = TRUE
       )
       # calculate Brier scores based on the 1st MCMC estimates
-      BrierScore0 <- Score(list("Bayesian Cox" = model_train),
+      BrierScore <- Score(list("Bayesian Cox" = model_train),
         formula = Surv(time, status) ~ 1,
         data = data_test, conf.int = FALSE,
-        metrics = "brier", summary = "ibs",
-        times = times
+        metrics = "brier", summary = c("ibs", "ipa"),
+        null.model = TRUE, times = times
       )$Brier$score
-      Brier <- as.matrix(BrierScore0[1:(nrow(BrierScore0) / 2), -c(1:2)])
+      BrierScore[is.na(BrierScore)] <- 0
+      # Brier scores of NULL.model do not change
+      Brier.null <- BrierScore[1:(nrow(BrierScore) / 2), -c(1:2)]
+      Brier <- BrierScore[-c(1:(nrow(BrierScore) / 2)), -c(1:2)]
+      
       # calculate Brier scores based on other MCMC estimates
       for (i in 2:nrow(betas)) {
         data_train$lp <- survObj$lp.all[, i] # lp_all_train[, i]
@@ -254,24 +263,33 @@ predict.BayesSurvive <- function(object, survObj.new, type = "brier",
         BrierScore <- Score(list("Bayesian Cox" = model_train),
           formula = Surv(time, status) ~ 1,
           data = data_test, conf.int = FALSE,
-          metrics = "brier", summary = "ibs",
-          null.model = FALSE, times = times
+          metrics = "brier", summary = c("ibs", "ipa"),
+          null.model = TRUE, times = times
         )$Brier$score
-        Brier <- Brier + as.matrix(BrierScore[, -c(1:2)])
+
+        BrierScore[is.na(BrierScore)] <- 0
+        Brier <- Brier + BrierScore[-c(1:(nrow(BrierScore) / 2)), -c(1:2)]
       }
-      BrierScore0 <- data.frame(BrierScore0)
-      BrierScore0[-c(1:(nrow(BrierScore0) / 2)), 3:4] <- Brier / nrow(betas)
-      Brier <- BrierScore0
+      Brier <- rbind(Brier.null, Brier / nrow(betas))
+      
       # extract IBS for Null model and the Bayesian Cox model
-      ibs[1, ] <- Brier$IBS[c(length(times), length(times) * 2)]
+      ibs[1, ] <- Brier$Brier[c(length(times), length(times) * 2)]
+      ibs[2, ] <- Brier$IBS[c(length(times), length(times) * 2)]
+      ibs[3, ] <- Brier$IPA[c(length(times), length(times) * 2)]
     }
-    # rownames(ibs) <- "IBS"
+    
+    #ibs <- rbind(rep(max(times), 2), ibs)
+    t0 <- round(max(times), digits = 1)
+    rownames(ibs) <- c(paste0("Brier(t=", t0, ")"), 
+                       paste0("IBS(t:0~", t0, ")"), 
+                       paste0("IPA(t=", t0, ")"))
     if (verbose) {
-      cat("                      IBS\n",
-        "  Null model          ", ibs[1, 1],
-        "\n  Bayesian Cox model  ", ibs[1, 2], "\n",
-        sep = ""
-      )
+      # cat("                      IBS\n",
+      #   "  Null model          ", ibs[1, 1],
+      #   "\n  Bayesian Cox model  ", ibs[1, 2], "\n",
+      #   sep = ""
+      # )
+      print(t(ibs))
     }
     invisible(Brier)
     # return(Brier)