+++ NEWS.md
@@ -1,7 +1,7 @@
 # BayesSurvive 0.0.2
 
 * Add vignette
-* Add function `coef.BayesSurvive()` to extract estimatiated coefficients and uncertainty
+* Add function `coef.BayesSurvive()` to extract estimated coefficients and uncertainty
 * Add function `VS()` to determine variable selection
 * Fix documentation issues